+++ AWSAuthSDK/Sources/AWSMobileClient/AWSMobileClientOperations/AWSAsyncOperations/GetTokensUserPool/FetchUserPoolTokensOperation.swift
@@ -31,7 +31,7 @@ class FetchUserPoolTokensOperation: AWSAsyncOperation {
     // Holds an instance of the user pool low level SDK
     //
     private let userPool: CognitoUserPoolBehavior
-    
+
     public weak var delegate: FetchUserPoolTokensDelegate?
     private var tokenFetchState: TokenFetchState = .notStarted
 
@@ -126,28 +126,22 @@ class FetchUserPoolTokensOperation: AWSAsyncOperation {
         }
 
         let user = self.userPool.getIdentityUser(username)
-        user.getSession().continueWith { [weak self] task in
-            guard let self = self else { return nil }
+        user.getUserPoolToken { [weak self] result in
+            guard let self = self else { return }
 
             guard
                 !self.isCancelled
             else {
                 AWSMobileClientLogging.verbose("\(self.identifier) Cancelled")
                 self.finish()
-                return nil
+                return
             }
 
-            if let tokens = task.result {
-                self.acceptEvent(.tokenFetched(tokens.mobileClientTokens))
-                return nil
-
-            } else if let error = task.error {
-
-                // If the error returned is `InvalidAuthenticationDelegate` it means that userPool
-                // client is trying to get authentication details, which means that the refresh
-                // token is expired. In this case, we let the delegate know that it require
-                // re-authentication and wait for the user to signIn.
-                if (error as NSError).isAuthDelegateError {
+            switch result {
+            case .success(let tokens):
+                self.acceptEvent(.tokenFetched(tokens))
+            case .failure(let error):
+                if (error as NSError).didTokenExpire {
                     AWSMobileClientLogging.verbose("\(self.identifier) Need re-authentication")
                     self.acceptEvent(.tokenExpired)
 
@@ -155,7 +149,6 @@ class FetchUserPoolTokensOperation: AWSAsyncOperation {
                     self.acceptEvent(.serviceError(error))
                 }
             }
-            return nil
         }
     }
 
@@ -170,7 +163,7 @@ class FetchUserPoolTokensOperation: AWSAsyncOperation {
         let error = AWSMobileClientError.notSignedIn(message: message)
         finishWithError(error: error)
     }
-    
+
     private func finishWithResult(tokens: Tokens) {
         AWSMobileClientLogging.verbose("\(self.identifier) Success")
         completion(tokens, nil)
@@ -185,10 +178,23 @@ class FetchUserPoolTokensOperation: AWSAsyncOperation {
 }
 
 extension NSError {
-    var isAuthDelegateError: Bool {
+
+    // The underlying SDK returns different errors when the user token is expired based on the signIn
+    // mechanism.
+    //
+    // In case if user is signedin via non-HostedUI method and if the error returned is
+    // `InvalidAuthenticationDelegate` it means that userPool client is trying to get
+    // authentication details, which means that the refresh token is expired.
+    // On the other hand if the user is signedin via HostedUI, the SDK returns
+    // `errorExpiredRefreshToken` when the token expire.
+    //
+    var didTokenExpire: Bool {
         let authCode = AWSCognitoIdentityClientErrorType.providerClientErrorInvalidAuthenticationDelegate.rawValue
         if domain == AWSCognitoIdentityProviderErrorDomain,
            code == authCode  {
+            return true
+        } else if domain == AWSCognitoAuthErrorDomain,
+                  code == AWSCognitoAuthClientErrorType.errorExpiredRefreshToken.rawValue {
             return true
         }
         return false